% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsmdn-package.R
\docType{data}
\name{allyarray}
\alias{allyarray}
\title{binary relational data}
\usage{
data(allyarray)
}
\description{
dataset on alliance relationships for 199 countries
from 1945 to 1989
}
\keyword{datasets}
