% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dutch}
\alias{dutch}
\title{Dutch classroom dataset by Knecht}
\format{4 adjacency matrices with 25 nodes}
\source{
\url{https://www.stats.ox.ac.uk/~snijders/siena/tutorial2010_data.htm}
}
\usage{
dutch
}
\description{
Dutch classroom dataset by Knecht
}
\keyword{datasets}
