% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fit_latent_space_network}
\alias{fit_latent_space_network}
\title{Fits a Latent Space Model for a Dynamic Network}
\usage{
fit_latent_space_network(Y, Y_miss, X_init, radii_init, beta_in, beta_out,
  nu_in, xi_in, nu_out, xi_out, tau_sq, tau_shape, tau_scale, sigma_sq,
  sigma_shape, sigma_scale, num_samples, num_burn_in, tune, tune_interval,
  step_size_x, step_size_beta, step_size_radii, seed)
}
\arguments{
\item{Y}{The adjacency matrix for each network,
shape (n_time_points, n_nodes, n_nodes)}
}
\description{
Fits a Latent Space Model for a Dynamic Network
}
