% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gammaAccProb}
\alias{gammaAccProb}
\title{Draw from inv gamma and adjust acceptance rate, for non-negative normal data}
\usage{
gammaAccProb(X, dims, Y, BIN, BOUT, alph, bta, ww, g2, g2new)
}
\arguments{
\item{X}{data cube}

\item{dims}{vector of dims}

\item{Y}{an n x n x T array of relational matrices, where the third dimension corresponds to different time periods}

\item{BIN}{betaIn value}

\item{BOUT}{betaOut value}

\item{ww}{vector of weights}

\item{g2}{current value}

\item{g2}{new value}
}
\value{
returns list of starting values:
\item{g2New}
\item{AccRate}
}
\description{
Draw from inv gamma and adjust acceptance rate, for non-negative normal data
}

