% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gammaAccProbGaussian}
\alias{gammaAccProbGaussian}
\title{Draw from Metropolis Hasting acceptance probability for Gamma^2, for continuous data}
\usage{
gammaAccProbGaussian(X, dims, Y, BIN, alpha, aa, bb, ww, g2, g2new, WL)
}
\arguments{
\item{X}{an n x p x T array of latent coordinates, where the second dimension is the number dimensions of the latent space, and the third is time}

\item{dims}{vector of dimensions of X}

\item{Y}{an n x n x T array of relational matrices, where the third dimension corresponds to different time periods}

\item{BIN}{betaIn value}

\item{ww}{vector of weights/radius}

\item{g2}{current value of Gamma^2}

\item{g2new}{proposed new value of Gamma^2}

\item{BOUT}{betaOut value}

\item{alph}{Shape parameter for Gamma^2 value}

\item{bta}{Scale parameter for Gamma^2 value}
}
\value{
returns new value for Gamma^2 and updated acceptance rate:
\item{g2New}{new g2 value}
\item{AccRate}{updated accRate}
}
\description{
Draw from Metropolis Hasting acceptance probability for Gamma^2, for continuous data
}
