% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStartingValues.R
\name{getStartingValues}
\alias{getStartingValues}
\title{Initial parameter values for lsmdn}
\usage{
getStartingValues( Y, p=2, family='binomial', llApprox=FALSE, missData=FALSE, N=1000, seed=6886)
}
\arguments{
\item{Y}{an n x n x T array of relational matrices, 
where the third dimension corresponds to different time periods.}

\item{W}{an n x n x T x p design array, where the third dimension corresponds to
time and the fourth to different covariates.}

\item{p}{number of latent dimensions}

\item{family}{type of model to run. Options include 'normal', 'nonNegNormal', 'poisson', 'binomial'.}

\item{llApprox}{logical indicating whether or not to utilize log-likelihood 
approximationg. Only available for binomial model types.}

\item{missData}{logical indicating whether to impute missing data.}

\item{N}{number of MCMC iterations.}

\item{seed}{random seed}

\item{s2Init}{starting value for s2}

\item{t2Init}{starting value for t2}

\item{xLatPos}{starting actor positions in latent space}

\item{betaInInit}{starting value for betaIn}

\item{betaOutInit}{starting value for betaOut}

\item{nuIn}{prior mean for betaIn}

\item{nuOut}{prior variance for betaOut}

\item{xiIn}{prior variance for betaIn}

\item{xiOut}{prior variance for betaOut}

\item{shapeT2}{shape parameter for t2}

\item{scaleT2}{scale parameter for t2}

\item{shapeS2}{shape parameter for s2}

\item{scaleS2}{shape parameter for s2}
}
\value{
returns list of starting values:
\item{w}{weights}
\item{X}{initial actor latent space positions calculated via GMDS}
\item{betaIn}{Starting value for importance of popularity}
\item{betaOut}{Relative value for importance of activity }
\item{nuIn}{Prior mean for beta-in}
\item{nuOut}{Prior mean for beta-out}
\item{xiIn}{Prior variance for beta-in}
\item{xiOut}{Prior varianerce for beta-out}
\item{t2}{Variance for latent positions in initial time period}
\item{shapeT2}{Shape parameter for prior distribution of tau2}
\item{scaleT2}{Scale parameter for prior distribution of tau2}
\item{s2}{Variance of movement in the latent space in subsequent periods}
\item{shapeS2}{Shape parameter for prior distribution of sigma2}
\item{scaleS2}{Scale parameter for prior distribution of sigma2}
if llApprox=TRUE, also returns
\item{dInMax}{add desc}
\item{dOutMax}{add desc}
\item{n0}{add desc}
\item{elOut}{add desc}
\item{elIn}{add desc}
\item{degree}{add desc}
\item{edgeList}{add desc}
}
\description{
Initial parameter values for lsmdn
}
