% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmds.R
\name{gmds}
\alias{gmds}
\title{Initial Latent Positions (GMDS, Sarkar and Moore, 2005)}
\usage{
gmds( Y )
}
\arguments{
\item{Y}{an n x n x T array of relational matrices, where the third dimension corresponds to different time periods}

\item{w}{weights or radii}

\item{p}{number of latent dimensions}

\item{family}{type of model to run. Options include 'normal', 'nonNegNormal', 'poisson', 'binomial'.}
}
\value{
Latent positions of actors based on Sarkar and Moore
}
\description{
Initial Latent Positions (GMDS, Sarkar and Moore, 2005)
}

