% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{imputeMissingGaussian}
\alias{imputeMissingGaussian}
\title{Impute missingness in Y}
\usage{
imputeMissingGaussian(X, dims, MM, Y, Ttt, BIN, BOUT, ww, g2)
}
\arguments{
\item{X}{data cube}

\item{dims}{vector of dims}

\item{MM}{}

\item{Y}{an n x n x T array of relational matrices, where the third dimension corresponds to different time periods}

\item{BIN}{betaIn value}

\item{BOUT}{betaOut value}

\item{ww}{vector of weights}
}
\value{
Y with imputed values
}
\description{
Impute missingness in Y
}

