% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{imputeMissingPoisson}
\alias{imputeMissingPoisson}
\title{Impute missingness in Y for Count Data}
\usage{
imputeMissingPoisson(X, dims, MM, Y, Ttt, BIN, alpha, ww, WL)
}
\arguments{
\item{X}{an n x p x T array of latent coordinates, where the second dimension is the number dimensions of the latent space, and the third is time}

\item{dims}{vector of dimensions of X}

\item{MM}{n x n x T array with 1s where Y is missing}

\item{Y}{an n x n x T array of relational matrices, where the third dimension corresponds to different time periods}

\item{Ttt}{time period value}

\item{BIN}{betaIn value}

\item{ww}{vector of weights/radius}

\item{BOUT}{betaOut value}
}
\value{
Y with imputed values
}
\description{
Impute missingness in Y for Count Data
}
