% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initBetaInOut.R
\name{initBetaInOut}
\alias{initBetaInOut}
\title{Initial betaIn and betaOUT; scale latent positions}
\usage{
gmds( Y )
}
\arguments{
\item{Y}{an n x n x T array of relational matrices, 
where the third dimension corresponds to different time periods.}

\item{xLatPos}{Latent space position of actors}

\item{p}{number of latent dimensions}

\item{w}{weights or radii}
}
\value{
returns list:
\item{betaInInit}{Initial value for $\beta_{IN}$}
\item{betaOutInit}{Initial value for $\beta_{OUT}$}
\item{xLatPos}{Scaled latent space}
}
\description{
Initial betaIn and betaOUT; scale latent positions
}

