% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{initialize_latent_positions}
\alias{initialize_latent_positions}
\title{Initializes Latent Postitions using GMDS (Sarkar and Moore, 2005)}
\usage{
initialize_latent_positions(Y, num_dimensions = 2, lambda = 10)
}
\description{
Initializes Latent Postitions using GMDS (Sarkar and Moore, 2005)
}
