% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{initializeBinom}
\alias{initializeBinom}
\title{initialize beta values for binomial data}
\usage{
initializeBinom(X, dims, Y, Xscale, BIN, BOUT, ww)
}
\arguments{
\item{X}{data cube}

\item{dims}{vector of dims}

\item{Y}{an n x n x T array of relational matrices, where the third dimension corresponds to different time periods}

\item{BIN}{betaIn value}

\item{BOUT}{betaOut value}

\item{ww}{vector of weights}
}
\description{
initialize beta values for binomial data
}

