% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{initializeBinomGrad}
\alias{initializeBinomGrad}
\title{initialize beta values for binomial data using a conjugate gradient method}
\usage{
initializeBinomGrad(X, dims, Y, Xscale, BIN, BOUT, ww)
}
\arguments{
\item{X}{n x p x T array of initial values for the latent space}

\item{dims}{vector of dimentions for X}

\item{Y}{an n x n x T array of relational matrices, where the third dimension corresponds to different time periods}

\item{Xscale}{1/n where n is the number of actors}

\item{BIN}{betaIn value}

\item{BOUT}{betaOut value}

\item{ww}{vector of radiuses/weights}
}
\description{
initialize beta values for binomial data using a conjugate gradient method
}
