% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initLogLikeApprox.R
\name{initLogLikeApprox}
\alias{initLogLikeApprox}
\title{Log-likelihood approximation subsampling}
\usage{
initWeights( Y, N )
}
\arguments{
\item{Y}{an n x n x T array of relational matrices, 
where the third dimension corresponds to different time periods.}

\item{n0}{n0}

\item{seed}{random seed}

\item{N}{number of MCMC iterations}
}
\value{
returns list of starting values for log likelihood approximation:
\item{dInMax}
\item{dOutMax}
\item{n0}
\item{elOut}
\item{elIn}
\item{subseq}
\item{degree}
\item{edgeList}
}
\description{
Log-likelihood approximation subsampling
}

