% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initNetMissVals.R
\name{initNetMissVals}
\alias{initNetMissVals}
\title{Initial values for missing entries}
\usage{
initNetMissVals( Y )
}
\arguments{
\item{Y}{an n x n x T array of relational matrices, where the third dimension corresponds to different time periods}

\item{missing}{List with information on missingness}

\item{seed}{random seed}
}
\value{
Inputted Y network with missingness filled in
}
\description{
Initial values for missing entries
}
