% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initWeights.R
\name{initWeights}
\alias{initWeights}
\title{Initial weights}
\usage{
initWeights( Y, N )
}
\arguments{
\item{Y}{an n x n x T array of relational matrices, 
where the third dimension corresponds to different time periods.}

\item{N}{number of MCMC iterations}
}
\value{
returns n x N matrix of weights or radii
}
\description{
Initial weights
}

