\name{lsmdn-package}
\alias{lsmdn-package}
\alias{lsmdn}
\docType{package}
\title{
\packageTitle{lsmdn}
}
\description{
\packageDescription{lsmdn}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{lsmdn}
\packageIndices{lsmdn}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{lsmdn}

Maintainer: \packageMaintainer{lsmdn}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
