% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lsmdn_log_likelihood}
\alias{lsmdn_log_likelihood}
\title{Calculates the log-likelihood of the latent space network model.}
\usage{
lsmdn_log_likelihood(Y, X, radii, beta_in, beta_out)
}
\description{
Calculates the log-likelihood of the latent space network model.
}
