% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lsmdn_predict_proba}
\alias{lsmdn_predict_proba}
\title{Calculate the probabilities of an edge formation according to the model}
\usage{
lsmdn_predict_proba(Y, X, radii, beta_in, beta_out)
}
\description{
Calculate the probabilities of an edge formation according to the model
}
