% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsmdn-package.R
\docType{data}
\name{matlCoopArrList}
\alias{matlCoopArrList}
\title{non-neg normal relational data}
\usage{
data(matlCoopArrList)
}
\description{
list of material cooperation arrays from
icews dataverse
}
\keyword{datasets}

