% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{postZeroProb}
\alias{postZeroProb}
\title{Posterior zero prob}
\usage{
postZeroProb(xi1, xi2, xj1, xj2, ss2, lam, pp0)
}
\arguments{
\item{xi1}{matrix}

\item{xi2}{matrix}

\item{xj1}{matrix}

\item{xj2}{matrix}

\item{ss2}{vector}
}
\value{
stuff
}
\description{
Posterior zero prob
}
