% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsmdn.R
\name{predict.lsmdn}
\alias{predict.lsmdn}
\title{Compute edge probabilities for the fitted adjacency matrix}
\usage{
\method{predict}{lsmdn}(model)
}
\description{
Compute edge probabilities for the fitted adjacency matrix
}
