% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{predictionBinomial}
\alias{predictionBinomial}
\title{get predictions for yhat}
\usage{
predictionBinomial(Ex, sig2, x1, x2, Bin, Bout, ww)
}
\arguments{
\item{Ex}{matrix}

\item{sig2}{vector}

\item{x1}{matrix}

\item{x2}{matrix}

\item{Bin}{betaIn vector}

\item{Bout}{betaOut vector}

\item{ww}{weights matrix}
}
\value{
yhat
}
\description{
get predictions for yhat
}

