% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{predictionGaussian}
\alias{predictionGaussian}
\title{get get predicted values for Y with continuous data}
\usage{
predictionGaussian(Ex, sig2, x1, x2, Bin, Bout, ww)
}
\arguments{
\item{Ex}{matrix of expected positions for X at time T}

\item{sig2}{vector of values for sigma^2 from different iterations}

\item{x1}{matrix of values for the first dimension of X from different iterations}

\item{x2}{matrix of values for the second dimention of X from different iterations}

\item{Bin}{vector of values for betaIn from different iterations}

\item{Bout}{vector of values for betaOut from different iterations}

\item{ww}{matrix of weights/radii from different iterations}
}
\value{
Matrix with expected values of Y
}
\description{
get get predicted values for Y with continuous data
}

