% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rdirichlet}
\alias{rdirichlet}
\title{Samples from a Dirichlet Distribution.}
\usage{
rdirichlet(num_samples, alphas, seed = 123L)
}
\description{
Note that you must change the seed per draw to get different values.
This is exposed for testing purposes and should not be used in R code.
}
