% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rinvgamma}
\alias{rinvgamma}
\title{Samples from an Inverse Gamma Distribution.}
\usage{
rinvgamma(num_samples, shape, scale = 1, seed = 123L)
}
\description{
Note that you must change the seed per draw to get different values.
This is exposed for testing purposes and should not be used in R code.
}
