% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_network.R
\name{rnorm_mixture}
\alias{rnorm_mixture}
\title{sample from a mixture of gaussians with equal mixture weights}
\usage{
rnorm_mixture(num_nodes, num_dimensions = 2, num_clusters = 5)
}
\description{
sample from a mixture of gaussians with equal mixture weights
}
