% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_network.R
\name{synthetic_network}
\alias{synthetic_network}
\title{Generate synthetic network data generated by a LSMDN model}
\usage{
synthetic_network(num_nodes = 100, num_time_steps = 10,
  num_dimensions = 2, beta_in = 1, beta_out = 2, seed = 42)
}
\description{
Generate synthetic network data generated by a LSMDN model
}
