% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{t2s2Parms}
\alias{t2s2Parms}
\title{generate shape and scale parameters to update t2 and s2 for all types of data}
\usage{
t2s2Parms(X, dims, thetaT, thetaS, phiT, phiS)
}
\arguments{
\item{X}{an n x p x T array of latent coordinates, where the second dimension is the number dimensions of the latent space, and the third is time}

\item{dims}{vector of dimensions of X}

\item{thetaT}{shape parameter for t}

\item{thetaS}{shape parameter for s}

\item{phiT}{scale parameter for t}

\item{phiS}{scale parameter for s}
}
\value{
returns list of:
\item{shapeT}{Shape Parameter for tau2}
\item{scaleT}{Scale Parameter for tau2}
\item{shapeS}{Shape Parameter for sigma2}
\item{scaleS}{Scale Parameter for sigma2}
}
\description{
generate shape and scale parameters to update t2 and s2 for all types of data
}

