% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{update_nnn}
\alias{update_nnn}
\title{update lat positions, beta params, and accrate for non-negative normal data}
\usage{
update_nnn(Xitm1, dims, tunex, Y, BIN, BOUT, tuneBIO, ww, t2, s2, g2, xiBin,
  xiBout, nuBin, nuBout, Cauchy, rnormsVec, rnormsBIO)
}
\arguments{
\item{Xitm1}{data cube}

\item{dims}{vector of dims}

\item{Y}{an n x n x T array of relational matrices, where the third dimension corresponds to different time periods}

\item{BIN}{betaIn value}

\item{BOUT}{betaOut value}

\item{ww}{vector of weights}

\item{rnormsVec}{vector}

\item{rnormsBIO}{vector}
}
\value{
returns list of:
\item{Xnew}
\item{BinNew}
\item{BoutNew}
\item{AccRate}
}
\description{
update lat positions, beta params, and accrate for non-negative normal data
}

