% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{updateBinom}
\alias{updateBinom}
\title{update lat positions, beta params, and accrate for binomial data}
\usage{
updateBinom(Xitm1, dims, tunex, Y, BIN, BOUT, tuneBIO, ww, t2, s2, xiBin,
  xiBout, nuBin, nuBout, Cauchy, rnormsVec, rnormsBIO)
}
\arguments{
\item{Xitm1}{an n x p x T array of the current latent coordinates, where the second dimension is the number dimensions of the latent space, and the third is time}

\item{dims}{vector of dimensions of Xitm1}

\item{tunex}{Variance of the normal random walk proposal for the latent space}

\item{Y}{an n x n x T array of relational matrices, where the third dimension corresponds to different time periods}

\item{BIN}{betaIn value}

\item{BOUT}{betaOut value}

\item{tuneBIO}{Variance of the normal random walk proposal for betaIn and betaOut}

\item{ww}{vector of radius/weights}

\item{t2}{variance of initial latent positions}

\item{s2}{variance of change in latent positions}

\item{xiBin}{mean of prior for betaIn}

\item{xiBout}{mean of prior for betaOut}

\item{nuBin}{variance of prior for betaIn}

\item{nuBout}{variance of prior for betaOut}

\item{Cauchy}{use a cauchy proposal or not}

\item{rnormsVec}{Vector of noise for random walks for X}

\item{rnormsBIO}{Vector of noise for random walks for betaIn/betaOut}
}
\value{
returns list of:
\item{Xnew}{New array of latent positions}
\item{BinNew}{New values for $\beta_{IN}$}
\item{BoutNew}{New values for $\beta_{OUT}}
\item{AccRate}{Updated acceptance rate}
}
\description{
update lat positions, beta params, and accrate for binomial data
}

