% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{updateBinomLogLikeApprox}
\alias{updateBinomLogLikeApprox}
\title{update lat positions, beta params, and accrate for binomial data using log-likelihood approximation}
\usage{
updateBinomLogLikeApprox(Xitm1, dims, tunex, Y, BIN, BOUT, tuneBIO, ww, t2, s2,
  xiBin, xiBout, nuBin, nuBout, Cauchy, rnormsVec, rnormsBIO, ELout, ELin,
  subseq, degr)
}
\arguments{
\item{Xitm1}{data cube}

\item{dims}{vector of dims}

\item{Y}{an n x n x T array of relational matrices, where the third dimension corresponds to different time periods}

\item{BIN}{betaIn value}

\item{BOUT}{betaOut value}

\item{ww}{vector of weights}

\item{rnormsVec}{vector}

\item{rnormsBIO}{vector}

\item{ELout}{array}

\item{ELin}{array}

\item{subseq}{matrix}

\item{degr}{array}
}
\value{
returns list of:
\item{Xnew}{add desc}
\item{BinNew}{add desc}
\item{BoutNew}{add desc}
\item{AccRate}{add desc}
}
\description{
update lat positions, beta params, and accrate for binomial data using log-likelihood approximation
}

