% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{wAccProb_llApprox}
\alias{wAccProb_llApprox}
\title{update weights and accprob for binomial data using log-likelihood approximation}
\usage{
wAccProb_llApprox(X, dims, Y, BIN, BOUT, tuneW, wwOld, wwNew, ELout, ELin,
  subseq, degr)
}
\arguments{
\item{X}{an n x p x T array of latent coordinates, where the second dimension is the number dimensions of the latent space, and the third is time}

\item{dims}{vector of dimensions of X}

\item{Y}{an n x n x T array of relational matrices, where the third dimension corresponds to different time periods}

\item{BIN}{betaIn value}

\item{BOUT}{betaOut value}

\item{tuneW}{variance of the proposal distribution for w}

\item{wwOld}{old vector of weights}

\item{wwNew}{new vector of weights}

\item{ELout}{array}

\item{ELin}{array}

\item{subseq}{matrix}

\item{degr}{array}
}
\value{
returns list of:
\item{wwNew}{add desc}
\item{AccRate}{add desc}
}
\description{
update weights and accprob for binomial data using log-likelihood approximation
}

