% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{wAccProbBinom}
\alias{wAccProbBinom}
\title{update weights and accprob for binary data}
\usage{
wAccProbBinom(X, dims, Y, BIN, BOUT, tuneW, wwOld, wwNew)
}
\arguments{
\item{X}{an n x p x T array of latent coordinates, where the second dimension is the number dimensions of the latent space, and the third is time}

\item{dims}{vector of dimensions of X}

\item{Y}{an n x n x T array of relational matrices, where the third dimension corresponds to different time periods}

\item{BIN}{betaIn value}

\item{BOUT}{betaOut value}

\item{tuneW}{variance of the proposal distribution for w}

\item{wwOld}{old vector of weights}

\item{wwNew}{new vector of weights}
}
\value{
returns list of:
\item{wwNew}{New weights/radius}
\item{AccRate}{Updated acceptance probability}
}
\description{
update weights and accprob for binary data
}

