% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{wAccProbGaussian}
\alias{wAccProbGaussian}
\title{update weights and accprob for non-negative normal data}
\usage{
wAccProbGaussian(X, dims, Y, BIN, BOUT, tuneW, wwOld, wwNew, g2)
}
\arguments{
\item{X}{data cube}

\item{dims}{vector of dims}

\item{Y}{an n x n x T array of relational matrices, where the third dimension corresponds to different time periods}

\item{BIN}{betaIn value}

\item{BOUT}{betaOut value}

\item{wwOld}{old vector of weights}

\item{wwNew}{new vector of weights}
}
\value{
returns list of:
\item{wwNew}{add desc}
\item{AccRate}{add desc}
}
\description{
update weights and accprob for non-negative normal data
}

