% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{wAccProbNonNegNormal}
\alias{wAccProbNonNegNormal}
\title{update weights and accprob for non-negative continuous normal data}
\usage{
wAccProbNonNegNormal(X, dims, Y, BIN, alpha, tuneW, wwOld, wwNew, g2, WL)
}
\arguments{
\item{X}{an n x p x T array of latent coordinates, where the second dimension is the number dimensions of the latent space, and the third is time}

\item{dims}{vector of dimensions of X}

\item{Y}{an n x n x T array of relational matrices, where the third dimension corresponds to different time periods}

\item{BIN}{betaIn value}

\item{tuneW}{variance of the proposal distribution for w}

\item{wwOld}{old vector of weights}

\item{wwNew}{new vector of weights}

\item{g2}{Variance of eta}

\item{BOUT}{betaOut value}
}
\value{
returns list of:
\item{wwNew}{New weights/radius}
\item{AccRate}{Updated acceptance probability}
}
\description{
update weights and accprob for non-negative continuous normal data
}

