% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStartingValues.R
\name{getStartingValues}
\alias{getStartingValues}
\title{Initial parameter values for lsmdn}
\usage{
getStartingValues( Y, p=2, family='binomial', llApprox=FALSE, missData=FALSE, N=1000, seed=6886)
}
\arguments{
\item{Y}{an n x n x T array of relational matrices, 
where the third dimension corresponds to different time periods.}

\item{p}{number of latent dimensions}

\item{family}{type of model to run. Options include 'normal', 'nonNegNormal', 'poisson', 'binomial'.}

\item{missData}{logical indicating whether to impute missing data.}

\item{s2Init}{starting value for s2}

\item{t2Init}{starting value for t2}

\item{xLatPos}{starting actor positions in latent space}

\item{betaInInit}{starting value for betaIn}

\item{betaOutInit}{starting value for betaOut}

\item{nuIn}{starting value for nuIn}

\item{nuOut}{starting value for nuOut}

\item{xiIn}{starting value for xiIn}

\item{xiOut}{starting value for xiOut}

\item{shapeT2}{shape parameter for t2}

\item{scaleT2}{scale parameter for t2}

\item{shapeS2}{shape parameter for s2}

\item{scaleS2}{shape parameter for s2}

\item{N}{number of MCMC iterations.}

\item{seed}{random seed}

\item{llApprox}{logical indicating whether or not to utilize log-likelihood 
approximationg. Only available for binomial model types.}
}
\value{
returns list of starting values:
\item{w}{weights}
\item{X}{initial actor latent space positions calculated via GMDS}
\item{betaIn}
\item{betaOut}
\item{nuIn}
\item{nuOut}
\item{xiIn}
\item{t2}
\item{shapeT2}
\item{scaleT2}
\item{s2}
\item{shapeS2}
\item{scaleS2}
if llApprox=TRUE, also returns
\item{dInMax}
\item{dOutMax}
\item{n0}
\item{elOut}
\item{elIn}
\item{degree}
\item{edgeList}
}
\description{
Initial parameter values for lsmdn
}

