% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsmdn-package.R, R/lsmdn.R, R/lsmdnSvelte.R
\docType{package}
\name{lsmdn}
\alias{lsmdn}
\alias{lsmdn-package}
\title{lsmdn - latent space model for dynamic networks}
\usage{
lsmdn( Y, p=2, family='binomial', llApprox=FALSE, missData=FALSE, N=1000, seed=6886)

lsmdn( Y, p=2, family='binomial', llApprox=FALSE, missData=FALSE, N=1000, seed=6886)
}
\arguments{
\item{Y}{an n x n x T array of relational matrices, 
where the third dimension corresponds to different time periods.}

\item{p}{number of latent dimensions}

\item{family}{type of model to run. Options include 'normal', 'nonNegNormal', 'poisson', 'binomial'.}

\item{llApprox}{logical indicating whether or not to utilize log-likelihood 
approximation. Only available for binomial model types.}

\item{missData}{logical indicating whether to impute missing data.}

\item{N}{number of MCMC iterations.}

\item{seed}{random seed}

\item{burnin}{Number of iterations to discard as burn-in}

\item{progressBar}{include progress bar to show mcmc status}

\item{saveResults}{save results as an rda file}

\item{savePoints}{chain intervals to save at}

\item{fileName}{"lsmdn.rda" or wtv you want, make sure to specify a path as well}

\item{tuneX}{variance of latent space proposal}

\item{tuneBIO}{variance of proposals for betaIn and betaOut}

\item{kappa}{variance of proposals for w}

\item{startVals}{Fitted result from previous model run.}

\item{odens}{thinning parameter}

\item{s2Init}{starting value for s2}

\item{t2Init}{starting value for t2}

\item{xLatPos}{starting actor positions in latent space}

\item{betaInInit}{starting value for betaIn}

\item{betaOutInit}{starting value for betaOut}

\item{nuIn}{prior mean of betaIn}

\item{nuOut}{prior mean of betaOut}

\item{xiIn}{prior variance of betaIn}

\item{xiOut}{prior variance of betaOut}

\item{shapeT2}{shape parameter for t2}

\item{scaleT2}{scale parameter for t2}

\item{shapeS2}{shape parameter for s2}

\item{scaleS2}{shape parameter for s2}

\item{shapeG2}{shape parameter for g2}

\item{scaleG2}{shape parameter for g2}

\item{odens}{How much to thin the posterior.}

\item{t2Init}{starting value for g2}

\item{Y}{an n x n x T array of relational matrices, 
where the third dimension corresponds to different time periods.}

\item{p}{number of latent dimensions}

\item{family}{type of model to run. Options include 'normal', 'nonNegNormal', 'poisson', 'binomial'.}

\item{llApprox}{logical indicating whether or not to utilize log-likelihood 
approximation. Only available for binomial model types.}

\item{missData}{logical indicating whether to impute missing data.}

\item{N}{number of MCMC iterations.}

\item{seed}{random seed}

\item{burnin}{Number of iterations to discard as burn-in}

\item{progressBar}{include progress bar to show mcmc status}

\item{saveResults}{save results as an rda file}

\item{savePoints}{chain intervals to save at}

\item{fileName}{"lsmdn.rda" or wtv you want, make sure to specify a path as well}

\item{tuneX}{variance of latent space proposal}

\item{tuneBIO}{variance of proposals for betaIn and betaOut}

\item{kappa}{variance of proposals for w}

\item{startVals}{Fitted result from previous model run.}

\item{odens}{How much to thin the posterior.}

\item{s2Init}{starting value for s2}

\item{t2Init}{starting value for t2}

\item{t2Init}{starting value for g2}

\item{xLatPos}{starting actor positions in latent space}

\item{betaInInit}{starting value for betaIn}

\item{betaOutInit}{starting value for betaOut}

\item{nuIn}{prior mean of betaIn}

\item{nuOut}{prior mean of betaOut}

\item{xiIn}{prior variance of betaIn}

\item{xiOut}{prior variance of betaOut}

\item{shapeT2}{shape parameter for t2}

\item{scaleT2}{scale parameter for t2}

\item{shapeS2}{shape parameter for s2}

\item{scaleS2}{shape parameter for s2}

\item{shapeG2}{shape parameter for g2}

\item{scaleG2}{shape parameter for g2}
}
\value{
returns list of starting values:
\item{w}{weights or radius for each actor}
\item{X}{actor latent space positions}
\item{betaIn}{relative value of popularity}
\item{betaOut}{relative value of activity}
\item{nuIn}{Prior mean for $\beta_{IN}$}
\item{nuOut}{Prior mean for $\beta_{OUT}$}
\item{xiIn}{Prior variance for $beta_{IN}$}
\item{xiOut}{Prior varianerce for $beta_{OUT}$}
\item{t2}{Variance for latent positions in initial time period}
\item{shapeT2}{Shape parameter for prior distribution of $\tau^2$}
\item{scaleT2}{Scale parameter for prior distribution of $\tau^2$}
\item{s2}{Variance of movement in the latent space in subsequent periods}
\item{shapeS2}{Shape parameter for prior distribution of $\sigma^2$}
\item{scaleS2}{Scale parameter for prior distribution of $\sigma^2$}
if llApprox=TRUE, also returns
\item{dInMax}{add desc}
\item{dOutMax}{add desc}
\item{n0}{add desc}
\item{elOut}{add desc}
\item{elIn}{add desc}
\item{degree}{add desc}
\item{edgeList}{add desc}

returns list of starting values:
\item{w}{weights or radius for each actor}
\item{X}{actor latent space positions}
\item{betaIn}{relative value of popularity}
\item{betaOut}{relative value of activity}
\item{nuIn}{Prior mean for $\beta_{IN}$}
\item{nuOut}{Prior mean for $\beta_{OUT}$}
\item{xiIn}{Prior variance for $beta_{IN}$}
\item{xiOut}{Prior varianerce for $beta_{OUT}$}
\item{t2}{Variance for latent positions in initial time period}
\item{shapeT2}{Shape parameter for prior distribution of $\tau^2$}
\item{scaleT2}{Scale parameter for prior distribution of $\tau^2$}
\item{s2}{Variance of movement in the latent space in subsequent periods}
\item{shapeS2}{Shape parameter for prior distribution of $\sigma^2$}
\item{scaleS2}{Scale parameter for prior distribution of $\sigma^2$}
if llApprox=TRUE, also returns
\item{dInMax}{add desc}
\item{dOutMax}{add desc}
\item{n0}{add desc}
\item{elOut}{add desc}
\item{elIn}{add desc}
\item{degree}{add desc}
\item{edgeList}{add desc}
}
\description{
lsmdn - latent space model for dynamic networks

LSMDN model fitting routine for longitudinal network data

LSMDN model fitting routine for longitudinal network data
}
\author{
Daniel Sewell <daniel-sewell@uiowa.edu>, Max Gallop <max.gallop@strath.ac.uk>,
Shahryar Minhas <shahryar.minhas@duke.edu>
Maintainer: Shahryar Minhas <shahryar.minhas@duke.edu>
}

