% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R, R/lsmdn.R
\docType{package}
\name{lsmdn}
\alias{lsmdn}
\alias{lsmdn-package}
\title{lsmdn - latent space model for dynamic networks}
\usage{
lsmdn( Y, p=2, family='binomial', llApprox=FALSE, missData=FALSE, N=1000, seed=6886)
}
\arguments{
\item{Y}{an n x n x T array of relational matrices, 
where the third dimension corresponds to different time periods.}

\item{p}{number of latent dimensions}

\item{family}{type of model to run. Options include 'normal', 'nonNegNormal', 'poisson', 'binomial'.}

\item{llApprox}{logical indicating whether or not to utilize log-likelihood 
approximationg. Only available for binomial model types.}

\item{missData}{logical indicating whether to impute missing data.}

\item{tuneX}{tuneX}

\item{tuneBIO}{tuneBIO}

\item{kappa}{kappa}

\item{burnin}{burnin}

\item{s2Init}{starting value for s2}

\item{t2Init}{starting value for t2}

\item{xLatPos}{starting actor positions in latent space}

\item{betaInInit}{starting value for betaIn}

\item{betaOutInit}{starting value for betaOut}

\item{nuIn}{starting value for nuIn}

\item{nuOut}{starting value for nuOut}

\item{xiIn}{starting value for xiIn}

\item{xiOut}{starting value for xiOut}

\item{shapeT2}{shape parameter for t2}

\item{scaleT2}{scale parameter for t2}

\item{shapeS2}{shape parameter for s2}

\item{scaleS2}{shape parameter for s2}

\item{N}{number of MCMC iterations.}

\item{seed}{random seed}

\item{startVals}{Fitted result from previous model run.}

\item{savePoints}{chain intervals to save at}

\item{fileName}{"lsmdn.rda" or wtv you want, make sure to specify a path as well}
}
\value{
returns list of starting values:
\item{w}{weights}
\item{X}{initial actor latent space positions calculated via GMDS}
\item{betaIn}{add desc}
\item{betaOut}{add desc}
\item{nuIn}{add desc}
\item{nuOut}{add desc}
\item{xiIn}{add desc}
\item{xiOut}{add desc}
\item{t2}{add desc}
\item{shapeT2}{add desc}
\item{scaleT2}{add desc}
\item{s2}{add desc}
\item{shapeS2}{add desc}
\item{scaleS2}{add desc}
if llApprox=TRUE, also returns
\item{dInMax}{add desc}
\item{dOutMax}{add desc}
\item{n0}{add desc}
\item{elOut}{add desc}
\item{elIn}{add desc}
\item{degree}{add desc}
\item{edgeList}{add desc}
}
\description{
lsmdn - latent space model for dynamic networks

LSMDN model fitting routine for longitudinal network data
}
\author{
Shahryar Minhas <shahryar.minhas@duke.edu>
}

