% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsmdn-package.R, R/lsmdn.R
\docType{package}
\name{lsmdn}
\alias{lsmdn}
\alias{lsmdn-package}
\alias{lsmdn-package}
\alias{lsmdn}
\title{lsmdn - latent space model for dynamic networks}
\usage{
lsmdn( Y, p=2, family='binomial', llApprox=FALSE, missData=FALSE, N=1000, seed=6886)
}
\arguments{
\item{Y}{an n x n x T array of relational matrices, 
where the third dimension corresponds to different time periods.}

\item{W}{an n x n x T x p design array, where the third dimension corresponds to
time and the fourth to different covariates.}

\item{p}{number of latent dimensions}

\item{family}{type of model to run. Options include 'normal', 'nonNegNormal', 'poisson', 'binomial'.}

\item{llApprox}{logical indicating whether or not to utilize log-likelihood 
approximation. Only available for binomial model types.}

\item{missData}{logical indicating whether to impute missing data.}

\item{N}{number of MCMC iterations.}

\item{seed}{random seed}

\item{burnin}{Number of iterations to discard as burn-in}

\item{odens}{thinning parameter}

\item{saveResults}{save results as an rda file}

\item{savePoints}{chain intervals to save at}

\item{startVals}{Fitted result from previous model run.}

\item{progressBar}{include progress bar to show mcmc status}

\item{fileName}{"lsmdn.rda" or wtv you want, make sure to specify a path as well}

\item{tuneX}{variance of latent space proposal}

\item{tuneBIO}{variance of proposals for betaIn and betaOut}

\item{tuneLAMBDA}{variance of proposals for lambda}

\item{kappa}{variance of proposals for w}
}
\value{
returns list of starting values:
\item{w}{weights or radius for each actor}
\item{X}{actor latent space positions}
\item{betaIn}{relative value of popularity}
\item{betaOut}{relative value of activity}
\item{nuIn}{Prior mean for beta-in}
\item{nuOut}{Prior mean for beta-out}
\item{xiIn}{Prior variance for beta-in}
\item{xiOut}{Prior varianerce for $beta-out}
\item{t2}{Variance for latent positions in initial time period}
\item{shapeT2}{Shape parameter for prior distribution of tau2}
\item{scaleT2}{Scale parameter for prior distribution of tau2}
\item{s2}{Variance of movement in the latent space in subsequent periods}
\item{shapeS2}{Shape parameter for prior distribution of sigma2}
\item{scaleS2}{Scale parameter for prior distribution of sigma2}
if llApprox=TRUE, also returns
\item{dInMax}{add desc}
\item{dOutMax}{add desc}
\item{n0}{add desc}
\item{elOut}{add desc}
\item{elIn}{add desc}
\item{degree}{add desc}
\item{edgeList}{add desc}
}
\description{
lsmdn - latent space model for dynamic networks

LSMDN model fitting routine for longitudinal network data
}
\author{
Daniel Sewell <daniel-sewell@uiowa.edu>, Max Gallop <max.gallop@strath.ac.uk>,
Shahryar Minhas <shahryar.minhas@duke.edu>
Maintainer: Shahryar Minhas <shahryar.minhas@duke.edu>
}
