% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{t2s2Parms}
\alias{t2s2Parms}
\title{update t2 and s2}
\usage{
t2s2Parms(X, dims, thetaT, thetaS, phiT, phiS)
}
\arguments{
\item{X}{data cube}

\item{dims}{vector of dims}

\item{thetaT}{shape parameter for t}

\item{thetaS}{shape parameter for s}

\item{phiT}{scale parameter for t}

\item{phiS}{scale parameter for s}
}
\value{
returns list of:
\item{shapeT}{add desc}
\item{scaleT}{add desc}
\item{shapeS}{add desc}
\item{scaleS}{add desc}
}
\description{
update t2 and s2
}

